Fix a bug parsing lines with only spaces
Adds a simple test file, broken-in.vtt, that starts with a single
space character, followed by a newline. In master this file panics
with an index out-of-bounds. This is fixed by checking both that
the line length is greater than zero, and that the resulting fields
length is also greater than zero.